requirements.txt
numpy==2.2.6
pygame==2.6.1
PyOpenGL==3.1.9
